requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
black==22.1.0
    # via -r requirements.in
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.0.3
    # via
    #   -r requirements.in
    #   black
idna==3.3
    # via requests
isort==5.10.1
    # via -r requirements.in
mypy-extensions==0.4.3
    # via black
numpy==1.22.2
    # via
    #   -r requirements.in
    #   pandas
pandas==1.4.1
    # via -r requirements.in
pathspec==0.9.0
    # via black
platformdirs==2.5.0
    # via black
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   pandas
pytz==2021.3
    # via pandas
requests==2.27.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
tomli==2.0.1
    # via black
typing-extensions==4.1.1
    # via black
urllib3==1.26.8
    # via requests
